+++ src/Bot.ts
@@ -1,9 +1,10 @@
 import {Client, Partials} from "discord.js";
 import mongoose from "mongoose";
 import {EnvSecrets} from "./EnvSecrets";
-import {CommonComponents, StrapListeners} from "./Listeners/_Listeners";
+import {StrapListeners} from "./Listeners/_Listeners";
 import {AIPool} from "./Types/AIPool";
 import {DbSeeder} from "./Database/Seeding/Seeder";
+import { CommonComponents } from "./CommonComponents";
 
 console.log("Bot is starting...");
 
@@ -12,17 +13,21 @@ const client = new Client({
     intents: ['DirectMessages', 'MessageContent', 'DirectMessageReactions', 'GuildMessages', 'GuildMessageReactions', 'Guilds', 'GuildMessageTyping', 'DirectMessageTyping']
 });
 
-mongoose.connect(EnvSecrets.getSecretOrThrow<string>('DB_CONNECTION_STRING'), {
+const dbPromise = mongoose.connect(EnvSecrets.getSecretOrThrow<string>('DB_CONNECTION_STRING'), {
     dbName: EnvSecrets.getSecretOrThrow<string>('DB_NAME'),
 }).then(async () => {
     console.log(`Connected to Database Server`);
     await DbSeeder.SeedDb();
 }).catch((err) => console.error(err));
 
+dbPromise.then((db) => {
+    console.log(db);
+})
+
 client.login(EnvSecrets.getSecretOrThrow<string>('TOKEN'));
 
 const ais: AIPool = new Map();
-const cc: CommonComponents = {ais, client, db: mongoose.connection};
+const cc: CommonComponents = {ais, client};
 
 // Trap client with listeners 
 StrapListeners(cc);
