+++ src/Listeners/TypingStart.ts
@@ -1,8 +1,7 @@
-import { channel } from "diagnostics_channel";
 import { Typing } from "discord.js";
 import { CheckAI } from "../Functions/CheckAI";
 import { CheckAllowedSource } from "../Functions/CheckAllowedSource";
-import { CommonComponents } from "./_Listeners";
+import { CommonComponents } from "../CommonComponents";
 
 export const TypingStartFunction = (typing: Typing, cc: CommonComponents) => {
     if (CheckAllowedSource(typing.channel.id, typing.guild?.id)) {
